+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [0.8.3]
 ### Improvements
+- New update lifecycle:
+    - now every package manager must provide the best upgrade order for all the selected packages ( can be disabled through the settings file **~/.config/bauh/config.yml** or the UI )
 - AUR
     - allowing the user to bypass checksum errors when installing / upgrading / downgrading packages